eature/AthreeSholdBeAssigned-part2

+++ database/migrations/2022_11_16_180552_create_threads_table.php
@@ -15,11 +15,12 @@ return new class extends Migration
     {
         Schema::create('threads', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('channel_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('channel_id');
             $table->string('title');
             $table->text('body');
             $table->timestamps();
+ 
         });
     }
